ported over the 3 existing pages to next w ssr, and dehydrated react query. pretty sure the theme cache is a bit buggy. will launch an investigation, opening PR and logging remaining work
